requirements.txt
pynput>=1.7.6
pyperclip>=1.8.2
anthropic>=0.18.0
pystray>=0.19.4
Pillow>=10.0.0
python-dotenv>=1.0.0
